+++ Order_Service/src/main/java/com/example/order/dto/InventoryProductResponse.java
@@ -8,9 +8,9 @@ public class InventoryProductResponse {
     private String productId;
     private String name;
     private Integer stock;
-    private Double price;
+    private Long price;
     @JsonProperty("discount_price")
-    private Double discountPrice;
+    private Long discountPrice;
 
     public String getProductId() {
         return productId;
@@ -36,19 +36,19 @@ public class InventoryProductResponse {
         this.stock = stock;
     }
 
-    public Double getPrice() {
+    public Long getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(Long price) {
         this.price = price;
     }
 
-    public Double getDiscountPrice() {
+    public Long getDiscountPrice() {
         return discountPrice;
     }
 
-    public void setDiscountPrice(Double discountPrice) {
+    public void setDiscountPrice(Long discountPrice) {
         this.discountPrice = discountPrice;
     }
 }
