+++ Order_Service/src/main/java/com/example/order/dto/OrderEventPayload.java
@@ -6,7 +6,7 @@ public class OrderEventPayload {
     private String productId;
     private String name;
     private int quantity;
-    private double totalPrice;
+    private Long totalPrice;
     private String status;
 
     public String getOrderId() {
@@ -49,11 +49,11 @@ public class OrderEventPayload {
         this.quantity = quantity;
     }
 
-    public double getTotalPrice() {
+    public Long getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(double totalPrice) {
+    public void setTotalPrice(Long totalPrice) {
         this.totalPrice = totalPrice;
     }
